catchError from server side done. Fixed searching.

+++ src/app/admin/movies-table/movies-table.component.ts
@@ -2,6 +2,9 @@ import {Component, OnInit} from '@angular/core';
 import {Movie} from '../../entities/Movie';
 import {MoviesService} from '../../services/movies.service';
 import {ActivatedRoute} from '@angular/router';
+import {ConfirmDialogComponent, ConfirmDialogData} from '../confirm-dialog/confirm-dialog.component';
+import {MatDialog} from '@angular/material/dialog';
+import {SnackbarService} from '../../services/snackbar.service';
 
 @Component({
     selector: 'app-movies-table',
@@ -10,9 +13,10 @@ import {ActivatedRoute} from '@angular/router';
 })
 export class MoviesTableComponent implements OnInit {
     movies: Movie[];
-    deletedMovie: Movie;
 
-    constructor(private moviesService: MoviesService, private route: ActivatedRoute) {
+
+    constructor(private moviesService: MoviesService, private route: ActivatedRoute, private dialog: MatDialog,
+                private snackbarService: SnackbarService) {
     }
 
     ngOnInit(): void {
@@ -26,10 +30,21 @@ export class MoviesTableComponent implements OnInit {
         );
     }
 
-    deleteMovie(movie: Movie): void {
-        this.deletedMovie = movie;
-        this.moviesService.deleteMovie(movie).subscribe(ok => {
-            this.movies = this.movies.filter(m => m !== movie);
+    deleteMovie(movie: Movie): any {
+        const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+            data: new ConfirmDialogData('Deleting movie', 'Do you really want to delete movie ' + movie.titleEN + ' '
+                + movie.secondTitleEN + '?')
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                this.moviesService.deleteMovie(movie).subscribe(ok => {
+                    if (ok) {
+                        this.movies = this.movies.filter(m => m.movieID !== movie.movieID);
+                        this.snackbarService.successMsg('Movie successfully deleted');
+                    }
+                });
+            }
         });
     }
